moving theme state management to layout component to fix issue where only half of components are updating to localstorage theme on deploy
