+++ src/components/Header/header.js
@@ -13,7 +13,7 @@ import MenuIcon from '../Icons/MenuIcon';
 
 class Header extends Component {
   render() {
-    const { pathname } = this.props;
+    const { pathname, updateTheme } = this.props;
 
     return (
       <VisualContextConsumer>
@@ -24,7 +24,7 @@ class Header extends Component {
           toggleNavMenu,
           colorMenuOpen,
           toggleColorMenu,
-          updateTheme,
+          // updateTheme,
         }) => (
           // <MenuContextConsumer>
           //   {({ menuOpen, toggleMenu }) => (
